#3, #10 (Rendertron): Refactor code, remove duplicate meta tags
