--- server/process.json
@@ -1,13 +1,0 @@
-    {
-      "name"        : "worker",
-      "script"      : "./index.js",
-      "instances"   : 1,
-      "merge_logs"  : true,
-      "log_date_format" : "YYYY-MM-DD HH:mm Z",
-      "watch": true,
-      "watch_options": {
-        "followSymlinks": true,
-        "usePolling"   : true,
-        "interval"    : 5
-      }
-    }
