+++ src/app/app.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { MetaTagsService } from './meta-tags.service';
 
 @Component({
   selector: 'app-root',
@@ -8,7 +7,4 @@ import { MetaTagsService } from './meta-tags.service';
 })
 export class AppComponent {
   title = 'Google Trends Watcher';
-  constructor(private metaTagsService: MetaTagsService) {
-    this.metaTagsService.addMetaTags('Home');
-  }
 }
