+++ src/app/dashboard/dashboard.component.ts
@@ -2,8 +2,7 @@ import { Component, OnInit } from '@angular/core';
 
 import { Category } from '../category';
 import { CategoryService } from '../category.service';
-
-import { MetaTagsService} from '../meta-tags.service';
+import { SeoService} from '../seo.service';
 
 @Component({
   selector: 'app-dashboard',
@@ -16,17 +15,16 @@ export class DashboardComponent implements OnInit {
 
   constructor(
     private categoryService: CategoryService,
-    private metaTagsService: MetaTagsService
+    private seo: SeoService
     ) { }
 
   ngOnInit() {
     this.getCategories();
 
     //SEO tags
-    this.metaTagsService.addMetaTags('Dashboard');
-    this.metaTagsService.setCanonicalURL();
-    this.metaTagsService.setPageTitle('Main Dashboard - Google Trends Watcher');
-    this.metaTagsService.setMetaDescription("Use the dashboard to find the latest trends to watch for in Google Trends");
+    this.seo.setCanonicalURL();
+    this.seo.setPageTitle('Main Dashboard - Google Trends Watcher');
+    this.seo.setMetaDescription('Use the dashboard to find the latest trends to watch for in Google Trends');
   }
 
   getCategories(): void {
